deleted old endpoint and url model
