+++ app/kontres/views/fetch_bookable_items.py
@@ -1,13 +1,13 @@
-from rest_framework.decorators import api_view
+from rest_framework import viewsets, status
 from rest_framework.response import Response
-from rest_framework import status
 from app.kontres.models.bookable_item import BookableItem
 from app.kontres.serializer.bookable_item_serializer import BookableItemSerializer
 
 
-@api_view(['GET'])
-def bookable_item_view(request):
-    if request.method == 'GET':
-        bookable_items = BookableItem.objects.all()
-        serializer = BookableItemSerializer(bookable_items, many=True)
+class BookableItemViewSet(viewsets.ViewSet):
+
+    # GET: Retrieve a list of all bookable items
+    def list(self, request):
+        queryset = BookableItem.objects.all()
+        serializer = BookableItemSerializer(queryset, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
